+++ CHANGES.md
@@ -3,6 +3,7 @@
 * Added camera roll mouse action. Can press R to reset roll.
 * Mouse strafing now remains active while at least one mouse button is still pressed, regardless of whether strafing key was released.
 * Addon now cooperates with "Lock camera to view" setting.
+* Improved ergonomics of mouse acceleration.
 
 # 1.0 (2019-10-01)
 * Initial release